+++ test/transform/resource/after-ecj/ValWeirdTypes.java
@@ -1,3 +1,4 @@
+import java.math.BigDecimal;
 import java.util.*;
 import lombok.val;
 public class ValWeirdTypes<Z> {
@@ -60,5 +61,11 @@ public class ValWeirdTypes<Z> {
     final @val int[] copy = intArray;
     final @val java.lang.Object[] single = multiDimArray[0];
     final @val int singleInt = copy[0];
+  }
+  public void arraysAsList() {
+    final @val java.util.List<java.lang.Class<? extends java.lang.Object>> x = Arrays.asList(String.class, BigDecimal.class);
+    for (final @val java.lang.Class<? extends java.lang.Object> y : x) 
+      {
+      }
   }
 }